+++ lib/features/appTerms/presentation/app_terms_screen.dart
@@ -11,19 +11,25 @@ import 'package:zachranobed/routes/app_router.gr.dart';
 import 'package:zachranobed/ui/widgets/button.dart';
 import 'package:zachranobed/ui/widgets/checkbox.dart';
 import 'package:zachranobed/ui/widgets/screen_scaffold.dart';
-import 'package:zachranobed/ui/widgets/snackbar/temporary_snackbar.dart';
 
+/// A screen that informs the user about application terms and conditions that need to be accepted.
 @RoutePage()
 class AppTermsScreen extends StatefulWidget {
-  const AppTermsScreen({Key? key}) : super(key: key);
+  /// A flag to determine if the user has no accepted version of the terms and conditions.
+  ///
+  /// If `true`, the screen will display the initial terms acceptance UI.
+  /// If `false`, the screen will display the UI for accepting a new version of the terms.
+  final bool hasNoAcceptedVersion;
+
+  /// Creates an [AppTermsScreen].
+  const AppTermsScreen({super.key, required this.hasNoAcceptedVersion});
 
   @override
   State<AppTermsScreen> createState() => _AppTermsScreen();
 }
 
 class _AppTermsScreen extends State<AppTermsScreen> {
-  final _setNewestAcceptedAppTermsUseCase =
-      GetIt.I<SetNewestAcceptedAppTermsUseCase>();
+  final _setNewestAcceptedAppTermsUseCase = GetIt.I<SetNewestAcceptedAppTermsUseCase>();
   var _areTermsAccepted = false;
 
   @override
@@ -35,22 +41,7 @@ class _AppTermsScreen extends State<AppTermsScreen> {
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-              SvgPicture.asset(
-                ZOStrings.certificationCheckPath,
-                width: 205,
-                height: 294,
-              ),
-              const SizedBox(height: GapSize.xl),
-              Text(
-                context.l10n!.appTermsTitle,
-                style: Theme.of(context).textTheme.titleLarge,
-              ),
-              const SizedBox(height: GapSize.xs),
-              Text(
-                context.l10n!.appTermsSubtitle,
-                textAlign: TextAlign.center,
-                style: Theme.of(context).textTheme.bodyLarge,
-              ),
+              ..._header(),
               const SizedBox(height: GapSize.m),
               ZOCheckbox.rich(
                 isChecked: _areTermsAccepted,
@@ -84,6 +75,7 @@ class _AppTermsScreen extends State<AppTermsScreen> {
               ZOButton(
                 text: context.l10n!.appTermsConfirm,
                 onPressed: _setNewestAcceptedAppTerms,
+                enabled: _areTermsAccepted,
               ),
             ],
           ),
@@ -92,17 +84,48 @@ class _AppTermsScreen extends State<AppTermsScreen> {
     );
   }
 
-  void _setNewestAcceptedAppTerms() {
-    if (mounted && _areTermsAccepted) {
-      _setNewestAcceptedAppTermsUseCase.invoke();
-      context.router.replace(const HomeRoute());
-    } else {
-      ScaffoldMessenger.of(context).showSnackBar(
-        ZOTemporarySnackBar(
-          backgroundColor: Colors.red,
-          message: context.l10n!.appTermsNotAcceptedError,
+  List<Widget> _header() {
+    if (widget.hasNoAcceptedVersion) {
+      return [
+        SvgPicture.asset(
+          ZOStrings.appTermsNotAcceptedPath,
         ),
-      );
+        const SizedBox(height: GapSize.xl),
+        Text(
+          context.l10n!.appTermsTitle,
+          style: Theme.of(context).textTheme.titleLarge,
+        ),
+        const SizedBox(height: GapSize.xs),
+        Text(
+          context.l10n!.appTermsSubtitle,
+          textAlign: TextAlign.center,
+          style: Theme.of(context).textTheme.bodyLarge,
+        ),
+      ];
+    } else {
+      return [
+        SvgPicture.asset(
+          ZOStrings.appTermsNewVersionPath,
+        ),
+        const SizedBox(height: GapSize.xl),
+        Text(
+          context.l10n!.appTermsNewVersionTitle,
+          style: Theme.of(context).textTheme.titleLarge,
+        ),
+        const SizedBox(height: GapSize.xs),
+        Text(
+          context.l10n!.appTermsNewVersionSubtitle,
+          textAlign: TextAlign.center,
+          style: Theme.of(context).textTheme.bodyLarge,
+        ),
+      ];
+    }
+  }
+
+  void _setNewestAcceptedAppTerms() async {
+    await _setNewestAcceptedAppTermsUseCase.invoke();
+    if (mounted) {
+      context.router.replace(const HomeRoute());
     }
   }